refactor(src/*): re postfix functions from now to txn
Lot's of the functions are named with a postfix now but really this often signifies it includes a
transaction, now relates to commting data for mutations in dgraph. So they are being renamed to
improve consistency with the dgraph library

BREAKING CHANGE: Function names have had the postfix changed. You will need to update your app.

fix #8